Implemented changes for name of ConfigLoader
