+++ src/main/java/interface_adapter/display_results_view/DisplayResultsController.java
@@ -11,6 +11,6 @@ public class DisplayResultsController {
     }
 
     public void execute(){
-        addLocationUseCaseInteractor.execute();
+//        addLocationUseCaseInteractor.execute();
     }
 }
